PYTHON/Gemini_1.5_Pro/Bridge/Round1/test_shape_renderer.py
from shape_renderer import Circle, Square, VectorRenderer, RasterRenderer

def test_circle_vector_rendering():
    # Arrange
    renderer = VectorRenderer()
    circle = Circle(renderer, 5)

    # Act
    result = circle.draw()  

    # Assert - Would usually capture output and check
    # assert result is None  # draw() doesn't return anything in this example

def test_circle_raster_rendering():
    # Arrange
    renderer = RasterRenderer()
    circle = Circle(renderer, 5)

    # Act
    result = circle.draw() 

    # Assert - Would usually capture output and check
    # assert result is None

def test_square_vector_rendering():
    # Arrange
    renderer = VectorRenderer()
    square = Square(renderer, 10)

    # Act
    result = square.draw() 

    # Assert - Would usually capture output and check
    # assert result is None

def test_square_raster_rendering():
    # Arrange
    renderer = RasterRenderer()
    square = Square(renderer, 10)

    # Act
    result = square.draw() 

    # Assert - Would usually capture output and check
    # assert result is None